pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.20.0",
    "pandas>=2.3.3",
    "streamlit-folium>=0.25.3",
    "streamlit>=1.50.0",
    "plotly>=6.3.1",
    "numpy>=2.3.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "replit-object-storage>=1.0.2",
]
